samples: wifi_provisioning_over_ble: Fix build
Fix build due to changed event callback signature in the upstream API.

+++ samples/wifi_provisioning_over_ble/src/app_config.h
@@ -14,5 +14,5 @@
 #endif /* CONFIG_WIFI_MGMT_RAW_SCAN_RESULTS_ONLY */
 
 static void start_dhcpv4_client(struct net_if *iface, void *user_data);
-static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				    struct net_if *iface);
